split main window components
